# 829 optimize testingstrategy