--- a/cpp/models/abm/config.h
+++ b/cpp/models/abm/config.h
@@ -1,0 +1,36 @@
+/*
+* Copyright (C) 2020-2024 MEmilio
+*
+* Authors: Daniel Abele
+*
+* Contact: Martin J. Kuehn <Martin.Kuehn@DLR.de>
+*
+* Licensed under the Apache License, Version 2.0 (the "License");
+* you may not use this file except in compliance with the License.
+* You may obtain a copy of the License at
+*
+*     http://www.apache.org/licenses/LICENSE-2.0
+*
+* Unless required by applicable law or agreed to in writing, software
+* distributed under the License is distributed on an "AS IS" BASIS,
+* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+* See the License for the specific language governing permissions and
+* limitations under the License.
+*/
+#ifndef MIO_ABM_CONFIG_H
+#define MIO_ABM_CONFIG_H
+
+namespace mio
+{
+namespace abm
+{
+
+/**
+ * Maximum number of age groups allowed in the model.
+ */
+const constexpr int MAX_NUM_AGE_GROUPS = 64;
+
+}
+} // namespace mio
+
+#endif