@dabele 2023-11-11
A few general notes
- map and set may be semantically convient in many cases, but are generally slow. unordered_set and map are much faster, but still slower than std::vector in many cases. these containers are generally only worth it for large data sets with sparse random accesses. Cutoff points need to be measured of course, but vector/array/CustomIndexArray is usually a good default choice, especially when the the data (AgeGroup, enums) can be used as an index. 
- map and unordered_map are also dangerous because `operator[]` adds an entry if it doesn't find it
- order of conditions matters
- we need to find ways to not do everything in every iteration/time step whenever possible. Even checking a single bool can be slow if it isn't in the CPU cache i.e. if it isn't used regularly.