@DavidKerkmann 2023-11-24
This together with #752 will also close #357.