@dabele 2023-12-08
Cleaned up the changes and fixed the unit tests.

Did better benchmark measurements to asses the changes. Uncertainty of ~2%.

baseline (current main): 8006 ms
with change 4 (vector instead of unordered_map for testing schemes): 6088 ms (~25%)
with change 4 and 2 (switch of ifs in testing strategy): 5640 ms (additional ~8 %)
with change 4, 2 and 3 (CustomIndexArray for GoToWork/GoToSchool paramater): 5638 ms (no measureable change)
with change 4, 2, 3 and 1 (bitset for agegroups in TestingCriteria with fixed number of age groups): 5262 ms (additional ~7%)

From our discussions changes 2,3,4 should be fine to add. Change 3 does not give performance, just cleaner code (I think). Change 1 adds some complexity to the model, we can discuss whether the performance gain is worth it. 

Reverted change 5 (only run TestingStrategy when migrating) as discussed. But I measured the performance for this change as well. Only testing persons that migrate to a different location gives 3889 ms (~27%). This measurement includes all the other changes above, I did not measure the change independently. So we probably should keep this in the discussion but in a separate Issue/MR.